move graphers, add ViewpsheetGrapher
