+++ src/editor/contexts/MusicContext.js
@@ -1,6 +1,6 @@
-import Grapher from "calchart/Grapher";
 import Song from "calchart/Song";
 import BaseContext from "editor/contexts/BaseContext";
+import Grapher from "graphers/Grapher";
 import { MusicContextMenus as menus } from "menus/EditorContextMenus";
 import AddSongPopup from "popups/AddSongPopup";
 import EditSongPopup from "popups/EditSongPopup";
